Refactor and add new functionalities to build game
The code now conforms to the Single Responsibility Principle for better clarity and maintainability. The desire to extraction of methods and variables has led to a refactoring process, which resulted in the addition of the IMiniGame interface, a new 'GameStatus' enumeration was implemented in GameManager and methods were organized in corresponding categories. The tutorial phase was revised and methods regarding score were introduced. Moreover, the 'AudioCtrl' class was renamed to 'AudioManager' and string parsing exceptions were added to improve error handling. Changes in 'Gacha' files are related to AudioManager rename.