Remove AI Behavior and Task related Scripts and Assets
The commit removes all AI Behavior and Task related scripts and asset files from the Scripts directory of MiniGames/Shoot. This includes deletion of AIBehavior, AIBehaviorEditor, AITaskManager and IAITask.cs files along with their corresponding asset and metadata files. This cleanup was necessary due to change in the artificial intelligence strategy of the game.